// src/pages/Board/react-beautiful-dnd.d.ts
export type Id = string;
export type DraggableId = Id;
export type DroppableId = Id;
// ...
export interface DraggableLocation {
  droppableId: DroppableId;
  index: number;
}